Refactor imports and update global variables for improved module organization; remove unused constants and functions
